compilation fix

+++ src/persistencer.rs
@@ -182,8 +182,8 @@ impl Persistencer for FilePersister {
         match self.lmdb {
             Some(ref lmdb) => {
                 //build and use iterator if database is non-empty
-                let count = lmdb.db_env.stat().unwrap().entries();
-                if (count > counter) {
+                let count = lmdb.db_env.stat().unwrap().entries() as u64;
+                if count > counter {
                     let tx = try!(lmdb.db_env.begin_ro_txn());
                     {
                         let mut cursor: RoCursor = try!(tx.open_ro_cursor(lmdb.db));
